Move the entirety of bags_common into the main common lib
