+++ bags/src/test/scala/uk/ac/wellcome/platform/archive/bags/async/fixtures/UpdateStoredManifestServiceFixture.scala
@@ -2,7 +2,7 @@ package uk.ac.wellcome.platform.archive.bags.async.fixtures
 
 import uk.ac.wellcome.fixtures.TestWith
 import uk.ac.wellcome.platform.archive.bags.async.services.UpdateStoredManifestService
-import uk.ac.wellcome.platform.archive.bags.fixtures.StorageManifestVHSFixture
+import uk.ac.wellcome.platform.archive.common.fixtures.StorageManifestVHSFixture
 import uk.ac.wellcome.storage.fixtures.LocalDynamoDb.Table
 import uk.ac.wellcome.storage.fixtures.S3.Bucket
