+++ bags_api/src/main/scala/uk/ac/wellcome/platform/storage/bags/api/models/DisplayBagManifest.scala
@@ -1,7 +1,7 @@
 package uk.ac.wellcome.platform.storage.bags.api.models
 
 import io.circe.generic.extras.JsonKey
-import uk.ac.wellcome.platform.archive.bags.common.models.FileManifest
+import uk.ac.wellcome.platform.archive.common.models.FileManifest
 
 case class DisplayBagManifest(
   checksumAlgorithm: String,
