+++ common/src/main/scala/uk/ac/wellcome/platform/archive/common/models/StorageManifest.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.platform.archive.bags.common.models
+package uk.ac.wellcome.platform.archive.common.models
 
 import java.time.Instant
 
@@ -7,7 +7,6 @@ import uk.ac.wellcome.platform.archive.common.models.bagit.{
   BagId,
   BagInfo
 }
-import uk.ac.wellcome.platform.archive.common.models.StorageSpace
 import uk.ac.wellcome.platform.archive.common.progress.models.StorageLocation
 
 case class ChecksumAlgorithm(value: String)
