+++ common/src/test/scala/uk/ac/wellcome/platform/archive/common/generators/StorageManifestGenerators.scala
@@ -1,21 +1,21 @@
-package uk.ac.wellcome.platform.archive.bags.generators
+package uk.ac.wellcome.platform.archive.common.generators
 
 import java.time.Instant
 
-import uk.ac.wellcome.platform.archive.common.generators.{
-  BagInfoGenerators,
-  StorageSpaceGenerators
-}
 import uk.ac.wellcome.platform.archive.common.models.bagit.{
   BagDigestFile,
   BagInfo,
   BagItemPath
 }
+import uk.ac.wellcome.platform.archive.common.models.{
+  ChecksumAlgorithm,
+  FileManifest,
+  StorageManifest
+}
 import uk.ac.wellcome.platform.archive.common.progress.models.{
   StandardStorageProvider,
   StorageLocation
 }
-import uk.ac.wellcome.platform.archive.bags.common.models._
 import uk.ac.wellcome.storage.ObjectLocation
 
 trait StorageManifestGenerators
